requirements.txt
asgiref==3.8.1
asttokens==3.0.0
cffi==1.17.1
cryptography==45.0.2
decorator==5.2.1
Django==5.2
django-environ==0.12.0
django-extensions==4.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
executing==2.2.0
grpcio==1.71.0
grpcio-tools==1.71.0
ipython==9.2.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
matplotlib-inline==0.1.7
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.51
protobuf==5.29.4
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.19.1
PyJWT==2.9.0
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.13.2
wcwidth==0.2.13
